Add files via upload
Imagine you want to predict the price of a house. To do this, we use a technique called linear regression. It's like finding a line that best fits the data we have about houses. This line helps us make predictions about the price of a house based on its features.

We start by collecting data about different houses. We look at features like the size of the house (in square feet), the number of floors, the number of bedrooms, and the year the house was built. We also know the actual prices of these houses.

Next, we split the data into two parts: a training set and a testing set. We use the training set to teach the computer how to predict house prices. We show it the features of the houses in the training set and the corresponding prices, and it learns to find the best line that fits the data.

Once the model has learned from the training set, we test it using the testing set. We give it the features of the houses in the testing set, and it predicts the prices based on what it has learned. We compare these predicted prices with the actual prices to see how accurate our model is.

To measure the accuracy, we use metrics like the Root Mean Squared Error (RMSE). The RMSE tells us how close the predicted prices are to the actual prices. A lower RMSE means our model is doing a better job of predicting house prices.

By using this project, we can get an estimate of how much a house might cost based on its size, number of floors, number of bedrooms, and the year it was built. It helps us make informed decisions when buying or selling a house.

Remember, this is a simplified description, and there are more complex techniques and factors involved in real-world house price prediction. But I hope this description helps anyone, including kids, understand the basic idea behind this project.